requirements.txt
pyobjc-core
pyobjc-framework-Cocoa
pyobjc-framework-ApplicationServices
pyobjc-framework-Quartz
